requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
albucore==0.0.24
    # via albumentations
albumentations==2.0.8
    # via -r requirements.in
annotated-types==0.7.0
    # via pydantic
bcrypt==5.0.0
    # via -r requirements.in
blinker==1.9.0
    # via flask
certifi==2022.12.7
    # via requests
charset-normalizer==2.1.1
    # via requests
click==8.3.1
    # via flask
colorama==0.4.6
    # via click
contourpy==1.3.2
    # via matplotlib
cycler==0.12.1
    # via matplotlib
filelock==3.20.0
    # via torch
flask==3.1.2
    # via
    #   -r requirements.in
    #   flask-cors
    #   flask-jwt-extended
    #   flask-sqlalchemy
flask-cors==6.0.2
    # via -r requirements.in
flask-jwt-extended==4.7.1
    # via -r requirements.in
flask-sqlalchemy==3.1.1
    # via -r requirements.in
fonttools==4.61.1
    # via matplotlib
fsspec==2025.12.0
    # via torch
greenlet==3.3.0
    # via sqlalchemy
idna==3.4
    # via requests
imageio==2.37.2
    # via scikit-image
importlib-resources==6.5.2
    # via nibabel
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via
    #   flask
    #   torch
kiwisolver==1.4.9
    # via matplotlib
lazy-loader==0.4
    # via scikit-image
markupsafe==2.1.5
    # via
    #   flask
    #   jinja2
    #   werkzeug
matplotlib==3.10.8
    # via ultralytics
mpmath==1.3.0
    # via sympy
networkx==3.4.2
    # via
    #   scikit-image
    #   torch
nibabel==5.3.3
    # via -r requirements.in
numpy==2.2.6
    # via
    #   albucore
    #   albumentations
    #   contourpy
    #   imageio
    #   matplotlib
    #   nibabel
    #   opencv-python
    #   opencv-python-headless
    #   pywavelets
    #   scikit-image
    #   scipy
    #   tifffile
    #   torchvision
    #   ultralytics
    #   ultralytics-thop
opencv-python==4.12.0.88
    # via
    #   -r requirements.in
    #   ultralytics
opencv-python-headless==4.12.0.88
    # via
    #   albucore
    #   albumentations
packaging==24.1
    # via
    #   lazy-loader
    #   matplotlib
    #   nibabel
    #   pipdeptree
    #   scikit-image
pillow==12.0.0
    # via
    #   imageio
    #   matplotlib
    #   scikit-image
    #   torchvision
    #   ultralytics
pip==25.3
    # via pipdeptree
pipdeptree==2.28.0
    # via -r requirements.in
polars==1.36.1
    # via ultralytics
polars-runtime-32==1.36.1
    # via polars
psutil==7.2.1
    # via ultralytics
pydantic==2.12.5
    # via albumentations
pydantic-core==2.41.5
    # via pydantic
pyjwt==2.10.1
    # via flask-jwt-extended
pymysql==1.1.2
    # via -r requirements.in
pyparsing==3.3.1
    # via matplotlib
python-dateutil==2.9.0.post0
    # via matplotlib
python-dotenv==1.2.1
    # via -r requirements.in
pywavelets==1.8.0
    # via -r requirements.in
pyyaml==6.0.3
    # via
    #   albumentations
    #   ultralytics
requests==2.28.1
    # via ultralytics
scikit-image==0.25.2
    # via -r requirements.in
scipy==1.15.3
    # via
    #   albumentations
    #   scikit-image
    #   ultralytics
simsimd==6.5.12
    # via albucore
six==1.17.0
    # via python-dateutil
sqlalchemy==2.0.45
    # via flask-sqlalchemy
stringzilla==4.6.0
    # via albucore
sympy==1.13.1
    # via torch
tifffile==2025.5.10
    # via scikit-image
torch==2.5.1+cu121
    # via
    #   -r requirements.in
    #   torchaudio
    #   torchvision
    #   ultralytics
    #   ultralytics-thop
torchaudio==2.5.1+cu121
    # via -r requirements.in
torchvision==0.20.1+cu121
    # via
    #   -r requirements.in
    #   ultralytics
typing-extensions==4.15.0
    # via
    #   nibabel
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   torch
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
ultralytics==8.3.248
    # via -r requirements.in
ultralytics-thop==2.0.18
    # via ultralytics
urllib3==1.26.13
    # via requests
werkzeug==3.1.4
    # via
    #   flask
    #   flask-cors
    #   flask-jwt-extended
cryptography